+++ src/main.rs
@@ -13,13 +13,15 @@ use futures::stream;
 
 mod api;
 mod assets;
-mod render;
+mod renderer;
 mod svg_theme;
 mod theme;
 
 use api::{RequestBody, RequestParams};
-use render::Render;
+use renderer::svg_renderer::SVGRenderer;
 use theme::Themes;
+
+use crate::renderer::gif_renderer::GIFRenderer;
 
 #[derive(Parser)]
 struct Opt {
@@ -32,7 +34,7 @@ async fn image(themes: &'static Themes, Query(req): Query<RequestParams>) -> imp
     Response::builder()
         .header(CONTENT_TYPE, "image/gif")
         .body(StreamBody::new(stream::iter(
-            Render::new_image(themes, req, render::RenderFormat::GIF).map(Ok::<_, Infallible>),
+            GIFRenderer::new_image(themes, req).map(Ok::<_, Infallible>),
         )))
         .unwrap()
 }
@@ -41,7 +43,7 @@ async fn game(themes: &'static Themes, Json(req): Json<RequestBody>) -> impl Int
     Response::builder()
         .header(CONTENT_TYPE, "image/gif")
         .body(StreamBody::new(stream::iter(
-            Render::new_animation(themes, req).map(Ok::<_, Infallible>),
+            GIFRenderer::new_animation(themes, req).map(Ok::<_, Infallible>),
         )))
         .unwrap()
 }
@@ -50,14 +52,11 @@ async fn example(themes: &'static Themes) -> impl IntoResponse {
     game(themes, Json(RequestBody::example())).await
 }
 
-async fn example_svg(
-    themes: &'static Themes,
-    Query(req): Query<RequestParams>,
-) -> impl IntoResponse {
+async fn example_svg(Query(req): Query<RequestParams>) -> impl IntoResponse {
     Response::builder()
         .header(CONTENT_TYPE, "image/svg+xml")
         .body(StreamBody::new(stream::iter(
-            Render::new_image(themes, req, render::RenderFormat::SVG).map(Ok::<_, Infallible>),
+            SVGRenderer::new_image(req).map(Ok::<_, Infallible>),
         )))
         .unwrap()
 }
@@ -72,7 +71,7 @@ async fn main() {
         .route("/image.gif", get(move |req| image(themes, req)))
         .route("/game.gif", post(move |req| game(themes, req)))
         .route("/example.gif", get(move || example(themes)))
-        .route("/example.svg", get(move |req| example_svg(themes, req)));
+        .route("/example.svg", get(move |req| example_svg(req)));
 
     axum::Server::bind(&opt.bind)
         .serve(app.into_make_service())
