+++ src/renderer/mod.rs
@@ -1,0 +1,3 @@
+pub mod gif_renderer;
+pub mod renderer;
+pub mod svg_renderer;
