+++ src/renderer/renderer.rs
@@ -1,0 +1,37 @@
+use shakmaty::{uci::Uci, Bitboard, Board};
+
+pub enum RenderState {
+    Preamble,
+    Frame(RenderFrame),
+    Complete,
+}
+
+#[derive(Default, Debug)]
+pub struct RenderFrame {
+    pub board: Board,
+    pub highlighted: Bitboard,
+    pub checked: Bitboard,
+    pub delay: Option<u16>,
+}
+
+impl RenderFrame {
+    pub fn diff(&self, prev: &RenderFrame) -> Bitboard {
+        (prev.checked ^ self.checked)
+            | (prev.highlighted ^ self.highlighted)
+            | (prev.board.white() ^ self.board.white())
+            | (prev.board.pawns() ^ self.board.pawns())
+            | (prev.board.knights() ^ self.board.knights())
+            | (prev.board.bishops() ^ self.board.bishops())
+            | (prev.board.rooks() ^ self.board.rooks())
+            | (prev.board.queens() ^ self.board.queens())
+            | (prev.board.kings() ^ self.board.kings())
+    }
+}
+
+pub fn highlight_uci(uci: Option<Uci>) -> Bitboard {
+    match uci {
+        Some(Uci::Normal { from, to, .. }) => Bitboard::from(from) | Bitboard::from(to),
+        Some(Uci::Put { to, .. }) => Bitboard::from(to),
+        _ => Bitboard::EMPTY,
+    }
+}
